requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0

# Data
pandas>=2.0.0
numpy>=1.24.0

# LLM
openai>=1.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
langsmith>=0.1.0

# Testing
pytest>=8.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.23.0

# Dev
black>=24.0.0
ruff>=0.3.0
mypy>=1.8.0
